node, store: Move migrating db schemas to ConnectionPool and the StoreBuilder
