+++ node/src/main.rs
@@ -204,7 +204,7 @@ async fn main() {
     let expensive_queries = read_expensive_queries().unwrap();
 
     let store_builder =
-        StoreBuilder::new(&logger, &node_id, &config, metrics_registry.cheap_clone());
+        StoreBuilder::new(&logger, &node_id, &config, metrics_registry.cheap_clone()).await;
 
     graph::spawn(
         futures::stream::FuturesOrdered::from_iter(stores_eth_networks.flatten().into_iter().map(
